+++ src/Traits/MakeImmutable.php
@@ -4,15 +4,8 @@ namespace Larapie\DataTransferObject\Traits;
 
 trait MakeImmutable
 {
-    public function &__get($name)
-    {
-        $value = $this->properties[$name]->getValue();
-
-        return $value;
-    }
-
     protected function determineImmutability()
     {
-        $this->setImmutable();
+        $this->setImmutable(true);
     }
 }